requirements.txt
Pillow
exifread

